fix: behaviour of running light complete
the big problem was that the direction proceess dipendent on the next_state while the output is actually the current_state. Therefore the FPGA does not shows the light until the end. But I still do not have the answer of how come the simulation behavior differs from the FPGA implementation.